# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Wikipedia Titles Query Service
  version: 1.0.2
  license:
    name: UNLICENSED
    identifier: UNLICENSED
servers:
  - url: http://127.0.0.1:32123
security: []
paths:
  /health:
    get:
      summary: Service health and DB metadata
      operationId: getHealth
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  api_version:
                    type: string
                  db_path:
                    type: string
                  db_size_bytes:
                    type: integer
                  db_mtime:
                    type: string
        "400":
          description: Invalid request
        "500":
          description: Internal error
  /v1/titles/query:
    post:
      summary: Execute constrained SELECT query
      operationId: postTitlesQuery
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [sql]
              properties:
                sql:
                  type: string
                params:
                  type: array
                  items: {}
                max_rows:
                  type: integer
      responses:
        "200":
          description: Query result
          content:
            application/json:
              schema:
                type: object
                properties:
                  columns:
                    type: array
                    items:
                      type: string
                  rows:
                    type: array
                    items:
                      type: array
                      items: {}
                  row_count:
                    type: integer
                  truncated:
                    type: boolean
        "400":
          description: Invalid request
        "403":
          description: Query rejected by SQL policy (non-SELECT, forbidden object, or multi-statement)
        "413":
          description: Request/response exceeds limits
        "500":
          description: Internal error